+++ modules/Student_Billing/StudentPayments.php
@@ -10,8 +10,7 @@ if ( empty( $_REQUEST['print_statements'] ) )
 	Search( 'student_id', issetVal( $extra ) );
 }
 
-if ( ! empty( $_REQUEST['values'] )
-	&& $_POST['values']
+if ( $_REQUEST['modfunc'] === 'save'
 	&& AllowEdit()
 	&& UserStudentID() )
 {
@@ -63,8 +62,8 @@ if ( ! empty( $_REQUEST['values'] )
 		}
 	}
 
-	// Unset values, month_values, day_values, year_values, billing_fees & redirect URL.
-	RedirectURL( [ 'values', 'month_values', 'day_values', 'year_values', 'billing_fees' ] );
+	// Unset modfunc, values, dates, billing_fees & redirect URL.
+	RedirectURL( [ 'modfunc', 'values', 'month_values', 'day_values', 'year_values', 'billing_fees' ] );
 }
 
 if ( $_REQUEST['modfunc'] === 'remove'
@@ -228,7 +227,9 @@ if ( UserStudentID()
 
 	if ( empty( $_REQUEST['print_statements'] ) )
 	{
-		echo '<form action="' . URLEscape( 'Modules.php?modname=' . $_REQUEST['modname'] . '&student_id=' . UserStudentID() ) . '" method="POST">';
+		echo '<form action="' . URLEscape(
+			'Modules.php?modname=' . $_REQUEST['modname'] . '&modfunc=save&student_id=' . UserStudentID()
+		) . '" method="POST">';
 
 		if ( AllowEdit() )
 		{